Added a dashboard displaying listeners over the past 24 hours
Added Highcharts assets.
Added a DataController to Listeners for returning listener data.

Added a Snapshot EntityRepository for fetching data.
This contains MySQL-specific code for the moment, which will need to be
ported to Doctrine Query Language 'DQL'. However, the date/time
functions used will need to be registered for DQL - difficult to find a
good source.